correction (if needed)
We apply a correction to the result in the following cases:

If one of the original numbers is positive and the other is negative.

If the first number is negative and the second is positive, only if the last difference is not zero.

If both original numbers are negative and the last difference is zero.